bug_fixer_agent/requirements.txt
google-genai>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pathspec>=0.11.0
requests>=2.31.0
gitpython>=3.1.40
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest
pytest-django